// src/features/favorite/types/videolist/VideoSearchConditionType.ts
export type VideoSearchConditionType = {
    selectedFavoriteVideoCategory: string;
    selectedFavoriteVideoViewStatus: string;
    selectedFavoriteVideoTag: string;
    selectedFavoriteVideoFavoriteLevel: string;
    selectedFavoriteVideoSortKey: string;
    selectedFavoriteVideoPage: string;
    selectedFavoriteVideoFolder: string;
    selectedFavoriteVideoMode: string;
};